+++ tencentcloud/gs/v20191118/errors.go
@@ -53,6 +53,9 @@ const (
 	// 会话未找到。
 	RESOURCENOTFOUND_SESSIONNOTFOUND = "ResourceNotFound.SessionNotFound"
 
+	// 机器还在初始化中。
+	RESOURCEUNAVAILABLE_INITIALIZATION = "ResourceUnavailable.Initialization"
+
 	// 退出游戏中。
 	UNSUPPORTEDOPERATION_STOPPING = "UnsupportedOperation.Stopping"
 )