+++ tencentcloud/sms/v20210111/models.go
@@ -545,14 +545,22 @@ func (r *DescribeSmsSignListResponse) FromJsonString(s string) error {
 type DescribeSmsTemplateListRequest struct {
 	*tchttp.BaseRequest
 
-	// 模板 ID 数组。
-	// <dx-alert infotype="notice" title="注意">默认数组长度最大100</dx-alert>
-	TemplateIdSet []*uint64 `json:"TemplateIdSet,omitempty" name:"TemplateIdSet"`
-
 	// 是否国际/港澳台短信：
 	// 0：表示国内短信。
 	// 1：表示国际/港澳台短信。
 	International *uint64 `json:"International,omitempty" name:"International"`
+
+	// 模板 ID 数组。数组为空时默认查询模板列表信息（仅允许主账号使用），请使用 Limit 和 Offset 字段设置查询范围。
+	// <dx-alert infotype="notice" title="注意">默认数组长度最大100</dx-alert>
+	TemplateIdSet []*uint64 `json:"TemplateIdSet,omitempty" name:"TemplateIdSet"`
+
+	// 最大上限，最多100。
+	// 注：默认为0，TemplateIdSet 为空时启用。
+	Limit *uint64 `json:"Limit,omitempty" name:"Limit"`
+
+	// 偏移量。
+	// 注：默认为0，TemplateIdSet 为空时启用。
+	Offset *uint64 `json:"Offset,omitempty" name:"Offset"`
 }
 
 func (r *DescribeSmsTemplateListRequest) ToJsonString() string {
@@ -567,8 +575,10 @@ func (r *DescribeSmsTemplateListRequest) FromJsonString(s string) error {
 	if err := json.Unmarshal([]byte(s), &f); err != nil {
 		return err
 	}
-	delete(f, "TemplateIdSet")
 	delete(f, "International")
+	delete(f, "TemplateIdSet")
+	delete(f, "Limit")
+	delete(f, "Offset")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeSmsTemplateListRequest has unknown keys!", "")
 	}